Switch to MinimalLambda host and refresh docs
