+++ src/AlexaVoxCraft.MinimalLambda/AlexaVoxCraft.MinimalLambda.csproj
@@ -1,0 +1,24 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <TargetFrameworks>net10.0;net8.0;net9.0</TargetFrameworks>
+        <LangVersion>default</LangVersion>
+    
+        <PackageId>AlexaVoxCraft.MinimalLambda</PackageId>
+        <Title>AlexaVoxCraft.MinimalLambda</Title>
+        <Description>Minimal API-style Lambda hosting for AlexaVoxCraft skills powered by MinimalLambda.</Description>
+        <PackageTags>alexa;lambda;aws;mediatR;voxcraft;hosting;middleware</PackageTags>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\AlexaVoxCraft.Lambda\AlexaVoxCraft.Lambda.csproj" />
+      <ProjectReference Include="..\AlexaVoxCraft.MediatR\AlexaVoxCraft.MediatR.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <PackageReference Include="MinimalLambda" Version="2.0.0-beta.3" />
+    </ItemGroup>
+
+</Project>
